Use iso3c for countries to avoid NAs in HDI
Looking at the dataset, I noticed that some countries had "NA" for the HDI because the names did not match exactly between the two original datasets. This can be fixed if we use the `countrycode` package to map the names to ISO3C codes. In the end the countryes in the `corruption` dataframe that do not have HDIs will be: "Taiwan" , "Kosovo" , "Korea (North)", "Somalia" (which aren't in the original data)

Also, there was a minor encoding issue that affected countries such as "Côte d'Ivoire", that was fixed using the "locale" option of `read_csv()`